// app/requests/_personal/getJob/getJobSuccess.type.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface GetJobSuccess {
  data: {
    id: number;
    selfEmployed: boolean;
    status: 1 | 2 | 3 | 4 | 5 | 6;
    place: {
      id: number;
      name: string;
      latitude: string;
      longitude: string;
      address_kladr: string;
      logo: string;
      region: {
        id: number;
        name: string;
        [k: string]: unknown;
      };
      brand: {
        id: number;
        name: string;
        logo: string;
        description: string;
        [k: string]: unknown;
      };
      [k: string]: unknown;
    };
    user: {
      id: number;
      name: string;
      phone: number;
      email: string;
      logo: string;
      roles: {
        id: number;
        name:
          | "admin"
          | "client"
          | "manager"
          | "recruiter"
          | "specialist"
          | "supervisor";
        [k: string]: unknown;
      }[];
      [k: string]: unknown;
    };
    radius: number;
    price: number;
    priceResult: number;
    income: number;
    forPay: number;
    viewActivity: {
      name: string;
      detailName: string;
      previewText: string;
      logo: string;
      [k: string]: unknown;
    };
    dateStart: string;
    dateEnd: string;
    needFoto: boolean;
    dateActivity: {
      id: number;
      timeStart: string;
      timeEnd: string;
      places: {
        id: number;
        name: string;
        latitude: string;
        longitude: string;
        address_kladr: string;
        logo?: string;
        region: {
          id: number;
          name: string;
          [k: string]: unknown;
        };
        brand: {
          id: number;
          name: string;
          logo?: string;
          description: string;
          [k: string]: unknown;
        };
        [k: string]: unknown;
      }[];
      [k: string]: unknown;
    }[];
    order: null | {
      id: number;
      selfEmployed: boolean;
      status: number;
      user: {
        id: number;
        phone: number;
        email: string;
        logo: string;
        roles: {
          id: number;
          name:
            | "admin"
            | "client"
            | "manager"
            | "recruiter"
            | "specialist"
            | "supervisor";
          [k: string]: unknown;
        }[];
        [k: string]: unknown;
      };
      [k: string]: unknown;
    };
    task: null | {
      id: number;
      selfEmployed: boolean;
      status: number;
      user: {
        id: number;
        phone: number;
        email: string;
        logo: string;
        roles: {
          id: number;
          name:
            | "admin"
            | "client"
            | "manager"
            | "recruiter"
            | "specialist"
            | "supervisor";
          [k: string]: unknown;
        }[];
        [k: string]: unknown;
      };
      [k: string]: unknown;
    };
    acceptingUser: {
      id: number;
      phone: number;
      email: string;
      logo: string;
      roles: {
        id: number;
        name:
          | "admin"
          | "client"
          | "manager"
          | "recruiter"
          | "specialist"
          | "supervisor";
        [k: string]: unknown;
      }[];
      radius: string;
      name: string;
      age: string;
      country: string;
      viewActivities: string[];
      status: 1 | 2 | 3 | 4 | 5 | 6;
      [k: string]: unknown;
    };
    reports: {
      dateStart: null | string;
      dateEnd: null | string;
      report: null | string[];
      dayActivityId: null | number;
      id: number;
      hours: null | string;
      status: number;
      [k: string]: unknown;
    }[];
    [k: string]: unknown;
  };
  [k: string]: unknown;
}